pyproject.toml
[tool.poetry]
name = "kipubot"
version = "0.2.0"
description = "Graphs for Friday bottle raffles."
authors = [
  "Mikael Siidorow <mikael.siidorow@aalto.fi>",
  "Kalle Ahlstrom <71292737+kqlski@users.noreply.github.com>",
]
license = "GPL-3.0"
readme = "README.md"


[tool.poe.tasks]
start = "python3 -m kipubot"
dev = "watchfiles 'poetry run  poe start' kipubot"
test = "pytest"
test_hot = "watchfiles 'pytest' kipubot tests"
lint = "ruff ."
type = "pytype ."
format = [{ cmd = "black ." }, { cmd = "ruff . --fix" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pandas = "^2.0.0"
matplotlib = "^3.7.1"
scipy = "^1.10.1"
python-dotenv = "^1.0.0"
python-telegram-bot = "^20.2"
psycopg = "^3.1.8"
openpyxl = "^3.1.2"
pytz = "^2023.3"
uncertainties = "^3.1.7"

[tool.poetry.group.dev.dependencies]
watchfiles = "^0.19.0"
ruff = "^0.0.261"
pytype = "^2023.4.11"
pytest = "^7.3.1"
black = "^23.3.0"
poethepoet = "^0.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py310"

select = [
  "A",
  "B",
  "C90",
  "DTZ",
  "E",
  "EM",
  "F",
  "FBT",
  "I",
  "ICN",
  "ISC",
  "N",
  "PLC",
  "PLE",
  "PLR",
  "PLW",
  "Q",
  "RUF",
  "S",
  "SIM",
  "TRY",
  "TID",
  "UP",
  "W",
  "YTT",
]
ignore = []

fixable = [
  "A",
  "B",
  "C90", # C
  "D",
  "E",
  "F",
  "G",
  "I",
  "N",
  "Q",
  "S",
  "W",
  "ANN",
  "ARG",
  "BLE",
  "COM",
  "DJ",
  "DTZ",
  "EM",
  "ERA",
  "EXE",
  "FBT",
  "ICN",
  "INP",
  "ISC",
  "NPY",
  "PD",
  "PGH",
  "PIE",
  "PL",
  "PT",
  "PTH",
  "PYI",
  "RET",
  "RSE",
  "RUF",
  "SIM",
  "SLF",
  "TCH",
  "TID",
  #"T",
  "TRY",
  "UP",
  "YTT",
]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
]

# Same as Black.

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"


[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]
